progress ...
on saving the message to the DB
