+++ README.md
@@ -4,3 +4,4 @@ Created with Create T3 App. Using Supabase as the backend.
 
 # Notes
 
+Create a Nextjs TypeScript React app that uses the libraries, tRPC, zod, prisma, tailwindcss. It will store a message, slugified version of the message and a sha1 hash of the message in PostgreSQL database.
