+++ next.config.mjs
@@ -15,4 +15,18 @@ function defineNextConfig(config) {
 export default defineNextConfig({
   reactStrictMode: true,
   swcMinify: true,
+  // webpack: config => {
+  //   config.experiments = { ...config.experiments, topLevelAwait: true };
+  //   return config;
+  // },
+  webpack: config => {
+    config.experiments = undefined;
+    config.experiments.topLevelAwait = true;
+    return config;
+  },
 });
+
+
+
+
+
