+++ package.json
@@ -16,12 +16,14 @@
     "@trpc/next": "^9.27.2",
     "@trpc/react": "^9.27.2",
     "@trpc/server": "^9.27.2",
+    "crypto-js": "^4.1.1",
     "next": "12.2.5",
     "next-auth": "^4.10.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.34.2",
     "react-query": "^3.39.2",
+    "slugify": "^1.6.5",
     "superjson": "^1.9.1",
     "zod": "^3.18.0"
   },
