+++ src/schema/msg.schema.ts
@@ -1,4 +1,4 @@
-import z from 'zod'
+import { z } from 'zod'
 
 export const createPostSchema = z.object({
     msg: z.string().max(256, 'Max title length is 356'),
@@ -11,4 +11,12 @@ export type CreateMsgInput = z.TypeOf<typeof createPostSchema>
 
 export const getSinglePostSchema = z.object({
     postId: z.string().uuid(),
-})
+});
+
+const Message = z.object({
+    msg: z.string(),
+    slug: z.string(),
+    hashedMessage: z.string(),
+});
+
+export type MessageData = z.infer<typeof Message>;
